+++ app/src/main/res/values/strings.xml
@@ -17,6 +17,7 @@
     <string name="navigate_up">Navigate up</string>
     <string name="optional_label">* optional</string>
     <string name="options_menu">Options menu</string>
+    <string name="save">Save</string>
     <string name="share">Share</string>
     <string name="sync_started">Synchronization started/enqueued</string>
